remove all builders filter option and do  other fixes
