+++ packages/react-app/components/Header.jsx
@@ -193,11 +193,6 @@ export default function Header({
                   </NextLink>
                 </MenuItem>
                 <MenuItem>
-                  <NextLink href="/admin/newBatches" exact>
-                    New Batches
-                  </NextLink>
-                </MenuItem>
-                <MenuItem>
                   <span onClick={handleStreamsUpdate}>Run stream indexer</span>
                 </MenuItem>
               </MenuList>
